+++ src/components/Settings/VueTorrent/DashboardItem.vue
@@ -6,15 +6,14 @@ defineEmits<{ update: [value: void] }>()
 
 <template>
   <tr class="table-row">
-    <td><v-icon icon="mdi-drag-vertical" /></td>
+    <td><v-icon icon="mdi-drag-vertical" class="dnd-handle" /></td>
     <td><v-btn density="compact" :icon="property.active ? 'mdi-checkbox-marked' : 'mdi-checkbox-blank-outline'" variant="flat" @click="$emit('update')" /></td>
     <td>{{ $t(`torrent.properties.${property.name}`) }}</td>
   </tr>
 </template>
 
 <style scoped>
-tr.table-row {
+td .dnd-handle {
   cursor: move;
-  margin-bottom: 12px !important;
 }
 </style>